支持授权隧道流量限制设置小数（如 0.5GB）
将 UserTunnel.flow 及相关 DTO 从 Long 改为 Double，
FlowController 和 TunnelAutoSwitchService 同步更新比较逻辑

+++ springboot-backend/src/main/java/com/admin/common/task/TunnelAutoSwitchService.java
@@ -168,7 +168,7 @@ public class TunnelAutoSwitchService {
     }
 
     private boolean isExhausted(UserTunnel ut) {
-        Long limit = ut.getFlow();
+        Double limit = ut.getFlow();
         if (limit == null || limit <= 0) return false;
         long used = (ut.getInFlow() != null ? ut.getInFlow() : 0L)
                   + (ut.getOutFlow() != null ? ut.getOutFlow() : 0L);
